ImGuiWindowing Fix nullptr on cursor destruction 02/23/2025 | 25w08a2

+++ TRAP/src/Layers/ImGui/ImGuiWindowing.cpp
@@ -1302,6 +1302,9 @@ void TRAP::INTERNAL::ImGuiWindowing::Shutdown()
 
 	for (ImGuiMouseCursor cursorN = 0; cursorN < ImGuiMouseCursor_COUNT; ++cursorN)
 	{
+		if(bd->MouseCursors[NumericCast<usize>(cursorN)] == nullptr)
+			continue;
+
 		WindowingAPI::DestroyCursor(bd->MouseCursors[NumericCast<usize>(cursorN)]);
 		bd->MouseCursors[NumericCast<usize>(cursorN)] = nullptr;
 	}
